Give t-top type to bindings in get-named-provides
This fixes the immediate issue that the last commit introduced and passes
tests. I can't figure out what cases this is relevant for; the failing test was
a REPL tests, which does do copying into the Globals list. This may build
larger static provides objects for things at the REPL.